chore: move code specific to controller to right directory
